+++ tests/test_integration.py
@@ -5,7 +5,7 @@ import pytest
 from BZI.pseudopots import FreeElectronModel
 from BZI.integration import *
 from BZI.symmetry import Lattice, make_ptvecs, make_rptvecs
-from BZI.sampling import make_cell_points
+from BZI.sampling import make_grid
 
 def test_rectangular():
     """This will test the rectangular methods of finding the total energy and the
@@ -13,7 +13,8 @@ def test_rectangular():
     """
 
     degree_list = range(1,5)
-    nvalence_list = [1,2]
+    nvalence_list = [1]
+    # Can't 
 
     for degree in degree_list:
         for nvalence in nvalence_list:                    
@@ -23,7 +24,6 @@ def test_rectangular():
             lat_centering = "prim"
             lattice = Lattice(lat_centering, lat_consts, lat_angles)
             
-            nvalence = 1
             free = FreeElectronModel(lattice, nvalence, degree)
             
             grid_consts = [40]*3
@@ -31,14 +31,20 @@ def test_rectangular():
             grid_centering = "prim"
             grid_vecs = make_ptvecs(grid_centering, grid_consts, grid_angles)
             rgrid_vecs = make_rptvecs(grid_vecs)
-            offset = np.sum(free.lattice.reciprocal_vectors, 1)/2
-            grid = make_cell_points(free.lattice.reciprocal_vectors, rgrid_vecs, offset)
+            offset = -np.dot(np.linalg.inv(rgrid_vecs), 
+                             np.dot(lattice.reciprocal_vectors, [.5]*3))
+            grid = make_grid(free.lattice.reciprocal_vectors, rgrid_vecs, offset)
 
             free.fermi_level = rectangular_fermi_level(free, grid)
             sphere_volume = 4./3*np.pi*free.fermi_level**(3./degree)
             occupied_volume = free.lattice.reciprocal_volume*free.nvalence_electrons/2
             fl_answer = (3*occupied_volume/(4*np.pi))**(degree/3.)
-    
+
+            print("degree ", degree)
+            print("ne ", nvalence)
+            print("shere volume ", sphere_volume)
+            print("occupied_volume ", occupied_volume)
+            
             assert np.isclose(sphere_volume, occupied_volume, 1e-1, 1e-1)
             assert np.isclose(free.fermi_level, fl_answer, 1e-2,1e-2)